Remove appoitmentWithTimeZone file
